fix filed handling
