+++ .gitignore
@@ -1,0 +1,3 @@
+input/
+__pychache__/
+output/
