+++ description.txt
@@ -1,6 +1,6 @@
 フィルターの仕方を指定してください。
 0:指定した時刻でフィルター
-1:指定した時刻でフィルターしたものを StatusCode と time-taken でフィルター
+1:指定した時刻でフィルターしたものを StatusCode と time-taken(平均+標準偏差) でフィルター
 2:指定した時刻でフィルターしたものを StatusCode でフィルター
 3:指定した時刻でフィルターしたものを time-taken(平均+標準偏差) でフィルター
 4:StatusCode でフィルター
