--- output/[filterted_by_StatusCode][filterted_by_term]test.config
@@ -1,1 +1,0 @@
-#Fields: date time s-ip cs-method cs-uri-stem cs-uri-query s-port cs-username c-ip cs(User-Agent) cs(Referrer) sc-status sc-substatus sc-win32-status time-taken2021-05-02 17:47:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 409 0 0 1002021-05-02 17:48:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 400 0 0 12002021-05-02 17:49:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 500 0 0 300#Fields: date time s-ip cs-method cs-uri-stem cs-uri-query s-port cs-username c-ip cs(User-Agent) cs(Referrer) sc-status sc-substatus sc-win32-status time-taken2021-05-02 17:47:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 409 0 0 1002021-05-02 17:48:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 400 0 0 12002021-05-02 17:49:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 500 0 0 300#Fields: date time s-ip cs-method cs-uri-stem cs-uri-query s-port cs-username c-ip cs(User-Agent) cs(Referrer) sc-status sc-substatus sc-win32-status time-taken2021-05-02 17:47:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 409 0 0 1002021-05-02 17:48:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 400 0 0 12002021-05-02 17:49:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 500 0 0 300#Fields: date time s-ip cs-method cs-uri-stem cs-uri-query s-port cs-username c-ip cs(User-Agent) cs(Referrer) sc-status sc-substatus sc-win32-status time-taken2021-05-02 17:47:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 409 0 0 1002021-05-02 17:48:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 400 0 0 12002021-05-02 17:49:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 500 0 0 300#Fields: date time s-ip cs-method cs-uri-stem cs-uri-query s-port cs-username c-ip cs(User-Agent) cs(Referrer) sc-status sc-substatus sc-win32-status time-taken2021-05-02 17:47:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 409 0 0 1002021-05-02 17:48:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 400 0 0 12002021-05-02 17:49:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 500 0 0 300
