+++ required.txt
@@ -1,0 +1,3 @@
+openpyxl
+pandas
+statistics
