+++ samplefile/test.log
@@ -1,0 +1,17 @@
+#Software: Microsoft Internet Information Services 10.0
+#Version: 1.0
+#Date: 2002-05-02 17:42:15  // when the first log file entry was recorded, which is when the entire log file was created
+#Fields: date time s-ip cs-method cs-uri-stem cs-uri-query s-port cs-username c-ip cs(User-Agent) cs(Referrer) sc-status sc-substatus sc-win32-status time-taken
+2021-05-02 17:42:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 450
+2021-05-02 17:45:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 670
+2021-05-02 17:46:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 500
+2021-05-02 17:47:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 409 0 0 100
+2021-05-02 17:48:15 172.22.255.255 POST /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 400 0 0 1200
+2021-05-02 17:49:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 500 0 0 300
+2021-05-02 17:50:15 172.22.255.255 OPTION /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 1100
+2021-05-02 17:52:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 1100
+2021-05-02 17:52:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 1100
+2021-05-02 17:52:15 172.22.255.255 POST /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 1100
+2021-05-02 17:52:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 1100
+2021-05-02 17:52:15 172.22.255.255 HEAD /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 1100
+2021-05-02 17:52:15 172.22.255.255 GET /images/picture.jpg - 20 testUser 172.30.255.255 Mozilla/4.0+(compatible;MSIE+5.5;+Windows+2000+Server) higeReferrer 200 0 0 1100
